requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
Django==4.0.4
django-debug-toolbar==4.4.6
djangorestframework==3.13.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.9
sqlparse==0.5.1
tzdata==2024.2
